requirements.txt
# Web scraping
playwright>=1.40.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23

# Data analysis
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
